Ajustes no controller Auth

+++ src/main/java/com/eduardopontes/romaneioapp/dto/JwtDto.java
@@ -13,5 +13,5 @@ public class JwtDto {
 
     private String prefix;
 
-    private Long expiry;
+    private Integer expiry;
 }
